fix dtype

+++ src/test/nn/moe/mlp_test.py
@@ -11,8 +11,10 @@ from ...utils import requires_gpu
 
 @requires_gpu
 def test_mlp():
-    mlp = MoEMLP(d_model=128, hidden_size=256, num_experts=2, init_device="cuda")
-    x = torch.randn(5, 128, device="cuda")
+    mlp = MoEMLP(
+        d_model=128, hidden_size=256, num_experts=2, init_device="cuda", dtype=torch.bfloat16
+    )
+    x = torch.randn(5, 128, device="cuda", dtype=torch.bfloat16)
     tokens_per_expert = torch.tensor([3, 2], device="cuda")
     out = mlp(x, tokens_per_expert)
     assert out.shape == (5, 128)
@@ -22,12 +24,16 @@ def run_mlp_with_expert_parallelism():
     ep_mesh = init_device_mesh(get_default_device().type, (dist.get_world_size(),))
 
     mlp = MoEMLP(
-        d_model=128, hidden_size=256, num_experts=dist.get_world_size() * 2, init_device="meta"
+        d_model=128,
+        hidden_size=256,
+        num_experts=dist.get_world_size() * 2,
+        init_device="meta",
+        dtype=torch.bfloat16,
     )
     mlp.apply_ep(ep_mesh)
     mlp.to_empty(device=get_default_device())
 
-    x = torch.randn(5, 128, device="cuda")
+    x = torch.randn(5, 128, device="cuda", dtype=torch.bfloat16)
     tokens_per_expert = torch.tensor([3, 2], device="cuda")
     out = mlp(x, tokens_per_expert)
